Rename and enable EXT_blend_func_extended WebGL draft extension https://bugs.webkit.org/show_bug.cgi?id=267465
Reviewed by Kimmo Kinnunen.

The EXT_blend_func_extended extension has been approved
and renamed to WEBGL_blend_func_extended by the WebGL WG.

Updated the implementation and the relevant tests.

* LayoutTests/TestExpectations:
* LayoutTests/platform/ios-simulator/webgl/webgl-draft-extensions-flag-default-expected.txt:
* LayoutTests/platform/ios-simulator/webgl/webgl-draft-extensions-flag-on-expected.txt:
* LayoutTests/webgl/1.0.x/conformance/extensions/webgl-blend-func-extended-expected.txt: Renamed from LayoutTests/webgl/1.0.x/conformance/extensions/ext-blend-func-extended-expected.txt.
* LayoutTests/webgl/1.0.x/conformance/extensions/webgl-blend-func-extended.html: Renamed from LayoutTests/webgl/1.0.x/conformance/extensions/ext-blend-func-extended.html.
* LayoutTests/webgl/2.0.y/conformance2/extensions/webgl-blend-func-extended-expected.txt: Renamed from LayoutTests/webgl/2.0.y/conformance2/extensions/ext-blend-func-extended-expected.txt.
* LayoutTests/webgl/2.0.y/conformance2/extensions/webgl-blend-func-extended.html: Renamed from LayoutTests/webgl/2.0.y/conformance2/extensions/ext-blend-func-extended.html.
* LayoutTests/webgl/resources/webgl-draft-extensions-flag.js:
* LayoutTests/webgl/resources/webgl_test_files/conformance/extensions/webgl-blend-func-extended.html: Renamed from LayoutTests/webgl/resources/webgl_test_files/conformance/extensions/ext-blend-func-extended.html.
* LayoutTests/webgl/resources/webgl_test_files/conformance2/extensions/webgl-blend-func-extended.html: Renamed from LayoutTests/webgl/resources/webgl_test_files/conformance2/extensions/ext-blend-func-extended.html.
* LayoutTests/webgl/resources/webgl_test_files/js/tests/webgl-blend-func-extended.js: Renamed from LayoutTests/webgl/resources/webgl_test_files/js/tests/ext-blend-func-extended.js.
(runTestNoExtension):
(runEnumTests):
(runQueryTests):
(runShaderTests):
(runMissingOutputsTests):
(runDrawBuffersLimitTests):
(runBlendingTests):
(runTest):
* LayoutTests/webgl/webgl-draft-extensions-flag-default-expected.txt:
* LayoutTests/webgl/webgl-draft-extensions-flag-off-expected.txt:
* LayoutTests/webgl/webgl-draft-extensions-flag-on-expected.txt:
* Source/WebCore/CMakeLists.txt:
* Source/WebCore/DerivedSources-input.xcfilelist:
* Source/WebCore/DerivedSources-output.xcfilelist:
* Source/WebCore/DerivedSources.make:
* Source/WebCore/Sources.txt:
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/bindings/js/JSDOMConvertWebGL.cpp:
(WebCore::convertToJSValue):
* Source/WebCore/html/canvas/WebGL2RenderingContext.cpp:
(WebCore::WebGL2RenderingContext::getExtension):
(WebCore::WebGL2RenderingContext::getSupportedExtensions):
* Source/WebCore/html/canvas/WebGLBlendFuncExtended.cpp: Renamed from Source/WebCore/html/canvas/EXTBlendFuncExtended.cpp.
(WebCore::WebGLBlendFuncExtended::WebGLBlendFuncExtended):
(WebCore::WebGLBlendFuncExtended::supported):
* Source/WebCore/html/canvas/WebGLBlendFuncExtended.h: Renamed from Source/WebCore/html/canvas/EXTBlendFuncExtended.h.
* Source/WebCore/html/canvas/WebGLBlendFuncExtended.idl: Renamed from Source/WebCore/html/canvas/EXTBlendFuncExtended.idl.
* Source/WebCore/html/canvas/WebGLExtension.h:
* Source/WebCore/html/canvas/WebGLRenderingContext.cpp:
(WebCore::WebGLRenderingContext::getExtension):
(WebCore::WebGLRenderingContext::getSupportedExtensions):
* Source/WebCore/html/canvas/WebGLRenderingContextBase.cpp:
(WebCore::WebGLRenderingContextBase::getParameter):
(WebCore::WebGLRenderingContextBase::extensionIsEnabled):
(WebCore::WebGLRenderingContextBase::loseExtensions):
* Source/WebCore/html/canvas/WebGLRenderingContextBase.h:

Canonical link: https://commits.webkit.org/273038@main
