--- LayoutTests/webgl/1.0.x/conformance/extensions/ext-blend-func-extended-expected.txt
@@ -1,4 +1,0 @@
-This test runs the WebGL Test listed below in an iframe and reports PASS or FAIL.
-
-Test: ../../../resources/webgl_test_files/conformance/extensions/ext-blend-func-extended.html
-[ PASS ] All tests passed
