+++ LayoutTests/webgl/1.0.x/conformance/extensions/webgl-blend-func-extended-expected.txt
@@ -1,0 +1,4 @@
+This test runs the WebGL Test listed below in an iframe and reports PASS or FAIL.
+
+Test: ../../../resources/webgl_test_files/conformance/extensions/webgl-blend-func-extended.html
+[ PASS ] All tests passed
