--- LayoutTests/webgl/2.0.y/conformance2/extensions/ext-blend-func-extended-expected.txt
@@ -1,4 +1,0 @@
-This test runs the WebGL Test listed below in an iframe and reports PASS or FAIL.
-
-Test: ../../../resources/webgl_test_files/conformance2/extensions/ext-blend-func-extended.html?webglVersion=2
-[ PASS ] All tests passed
