+++ LayoutTests/webgl/2.0.y/conformance2/extensions/webgl-blend-func-extended.html
@@ -1,0 +1,18 @@
+<!-- This file is auto-generated by Tools/Scripts/update-webgl-conformance-tests. DO NOT EDIT -->
+<!DOCTYPE html>
+<html>
+<head>
+<meta charset="utf-8">
+<title>WebGL Conformance Test Wrapper for webgl-blend-func-extended.html</title>
+<script type="text/javascript" src="../../../resources/js-test-pre.js"></script>
+<script type="text/javascript" src="../../../resources/webkit-webgl-test-harness.js"></script>
+</head>
+<body>
+<p>This test runs the WebGL Test listed below in an iframe and reports PASS or FAIL.</p>
+Test: <a href="../../../resources/webgl_test_files/conformance2/extensions/webgl-blend-func-extended.html?webglVersion=2">../../../resources/webgl_test_files/conformance2/extensions/webgl-blend-func-extended.html?webglVersion=2</a>
+<div id="iframe">
+<iframe src="../../../resources/webgl_test_files/conformance2/extensions/webgl-blend-func-extended.html?quiet=1&webglVersion=2" width="800" height="600"></iframe>
+</div>
+<div id="result"></div>
+</body>
+</html>
