--- LayoutTests/webgl/resources/webgl_test_files/conformance2/extensions/ext-blend-func-extended.html
@@ -1,26 +1,0 @@
-<!--
-Copyright (c) 2023 The Khronos Group Inc.
-Use of this source code is governed by an MIT-style license that can be
-found in the LICENSE.txt file.
--->
-
-<!DOCTYPE html>
-<html>
-<head>
-<meta charset="utf-8">
-<title>WebGL 2.0 EXT_blend_func_extended Conformance Tests</title>
-<link rel="stylesheet" href="../../resources/js-test-style.css"/>
-<script src="../../js/js-test-pre.js"></script>
-<script src="../../js/webgl-test-utils.js"></script>
-</head>
-<body>
-<canvas width="32" height="32" id="c"></canvas>
-<div id="description"></div>
-<div id="console"></div>
-<script>
-var contextVersion = 2;
-</script>
-<script src="../../js/tests/ext-blend-func-extended.js"></script>
-<script src="../../js/js-test-post.js"></script>
-</body>
-</html>
