+++ LayoutTests/webgl/webgl-draft-extensions-flag-off-expected.txt
@@ -2,12 +2,10 @@ This test outputs which WebGL draft extensions are available.
 
 On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".
 
-TEST COMPLETE: 11 PASS, 0 FAIL
+TEST COMPLETE: 9 PASS, 0 FAIL
 
 
-PASS webgl:EXT_blend_func_extended: Not supported
 PASS webgl:EXT_texture_mirror_clamp_to_edge: Not supported
-PASS webgl2:EXT_blend_func_extended: Not supported
 PASS webgl2:EXT_render_snorm: Not supported
 PASS webgl2:EXT_texture_mirror_clamp_to_edge: Not supported
 PASS webgl2:OES_sample_variables: Not supported
