+++ LayoutTests/webgl/webgl-draft-extensions-flag-on-expected.txt
@@ -2,15 +2,11 @@ This test outputs which WebGL draft extensions are available.
 
 On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".
 
-TEST COMPLETE: 21 PASS, 0 FAIL
+TEST COMPLETE: 17 PASS, 0 FAIL
 
 
-PASS webgl:EXT_blend_func_extended: Supported
-PASS webgl:EXT_blend_func_extended: Has object.
 PASS webgl:EXT_texture_mirror_clamp_to_edge: Supported
 PASS webgl:EXT_texture_mirror_clamp_to_edge: Has object.
-PASS webgl2:EXT_blend_func_extended: Supported
-PASS webgl2:EXT_blend_func_extended: Has object.
 PASS webgl2:EXT_render_snorm: Supported
 PASS webgl2:EXT_render_snorm: Has object.
 PASS webgl2:EXT_texture_mirror_clamp_to_edge: Supported
