--- Source/WebCore/html/canvas/EXTBlendFuncExtended.idl
@@ -1,37 +1,0 @@
-/*
- * Copyright 2023 The Chromium Authors. All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions
- * are met:
- *
- * 1.  Redistributions of source code must retain the above copyright
- *     notice, this list of conditions and the following disclaimer.
- * 2.  Redistributions in binary form must reproduce the above copyright
- *     notice, this list of conditions and the following disclaimer in the
- *     documentation and/or other materials provided with the distribution.
- *
- * THIS SOFTWARE IS PROVIDED BY APPLE AND ITS CONTRIBUTORS "AS IS" AND ANY
- * EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
- * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL APPLE OR ITS CONTRIBUTORS BE LIABLE FOR ANY
- * DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
- * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
- * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
- * ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF
- * THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-[
-    LegacyNoInterfaceObject,
-    Conditional=WEBGL,
-    GenerateIsReachable,
-] interface EXTBlendFuncExtended {
-    const unsigned long SRC1_COLOR_EXT           = 0x88F9;
-    const unsigned long SRC1_ALPHA_EXT           = 0x8589;
-    const unsigned long ONE_MINUS_SRC1_COLOR_EXT = 0x88FA;
-    const unsigned long ONE_MINUS_SRC1_ALPHA_EXT = 0x88FB;
-
-    const unsigned long MAX_DUAL_SOURCE_DRAW_BUFFERS_EXT = 0x88FC;
-};
